Improve separation of concerns of match model, orm and repository
